+++ src/load_inputs/load_resources_data.jl
@@ -414,6 +414,38 @@ function check_maintenance_applicability(r::AbstractResource)
     return ErrorMsg.(error_strings)
 end
 
+function check_fusion_applicability(r::AbstractResource)
+    applicable_resources = Thermal
+
+    not_set = default_zero
+    value = get(r, :fusion, not_set)
+
+    error_strings = String[]
+
+    if value == not_set
+        # not FUSION so the rest is not applicable
+        return error_strings
+    end
+
+    if !isa(r, applicable_resources) && value > 0
+        e = string("Resource ", resource_name(r), " has :fusion = ", value, ".\n",
+            "This setting is valid only for resources where the type is \n",
+            "one of $applicable_resources. \n"
+        )
+        push!(error_strings, e)
+    end
+
+    model = get(r, :model, not_set)
+    if model == 2
+        e = string("Resource ", resource_name(r), " has :fusion = ", value, ".\n",
+            "This is valid only for resources with unit commitment (:model = 1);\n",
+            "this has :model = 2.")
+        push!(error_strings, e)
+    end
+
+    return ErrorMsg.(error_strings)
+end
+
 function check_retrofit_resource(r::AbstractResource)
     error_strings = String[]
 
@@ -467,6 +499,7 @@ function check_resource(r::AbstractResource)
     e = []
     e = [e; check_LDS_applicability(r)]
     e = [e; check_maintenance_applicability(r)]
+    e = [e; check_fusion_applicability(r)]
     e = [e; check_mustrun_reserve_contribution(r)]
     e = [e; check_retrofit_resource(r)]
     e = [e; check_qualified_hydrogen_supply(r)]