+++ src/model/resources/fusion/fusion_maintenance.jl
@@ -1,0 +1,33 @@
+# higher level function
+function fusion_maintenance_adjust_parasitic_power!(EP, rs::Vector{<:AbstractResource})
+    @debug "Maintenance modifies fusion parasitic power"
+
+    FUSION = ids_with(rs, fusion)
+    MAINTENANCE = ids_with_maintenance(rs)
+
+    for y in intersect(FUSION, MAINTENANCE)
+        resource_component = resource_name(rs[y])
+        reactor = FusionReactorData(rs, y)
+        _fusion_maintenance_parasitic_power_adjustment!(EP, resource_component, reactor)
+    end
+end
+
+# lower level function
+function _fusion_maintenance_parasitic_power_adjustment!(
+        EP::Model, resource_component, reactor::FusionReactorData)
+    passive = reactor.parasitic_passive_fraction
+    reduction_factor = reactor.maintenance_remaining_parasitic_power_fraction
+    η = reactor.eff_down
+    cap_size = reactor.component_size
+
+    reduction = passive * (1 - reduction_factor)
+
+    get_from_model(f::Function) = EP[Symbol(f(resource_component))]
+
+    ePassiveParasitic = get_from_model(fusion_parasitic_passive_name)
+    vMDOWN = get_from_model(maintenance_down_name)
+
+    eReduction = -reduction * η * cap_size * vMDOWN
+
+    add_similar_to_expression!(ePassiveParasitic, eReduction)
+end